backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-fastapi-production"
version = "1.0.0"
description = "Production-grade FastAPI application with LangChain integration"
# readme = "README.md"
license = "MIT"
authors = [{ name = "Harmeet Singh", email = "harmeetsinghfbd@gmail.com" }]
maintainers = [{ name = "Harmeet Singh", email = "harmeetsinghfbd@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
keywords = ["fastapi", "langchain", "ai", "api", "async"]
requires-python = ">=3.12,<3.13"
dependencies = [
    # --- CORE FRAMEWORK (Relaxed Upper Bounds) ---
    "fastapi[standard]==0.124.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
    "uvloop>=0.21.0,<1.0.0",
    "pydantic>=2.9.0,<3.0.0",
    "pydantic-settings>=2.5.0,<3.0.0", # Latest major is 2.11.0 (Sep 24, 2025)
    "httpx>=0.27.0,<1.0.0",
    "psutil==7.1.0",
    "loguru==0.7.3",
    "nanoid==2.0.0",
    "redis[hiredis]>=5.1.0,<6.0.0",
    "motor==3.7.1", # Keep tight for C-lib stability
    "prometheus-client>=0.23.1",
    "orjson==3.10.1",
    "beanie>=2.0.1",
    "httptools>=0.7.1",
    "pymongo>=4.11",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "bcrypt>=4.0.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    # --- TESTING (Updated) ---
    "pytest>=8.2.2,<9.0.0", # Latest is 8.2.2 (Sep 24, 2025) - Relaxed to <9.0.0
    "pytest-asyncio>=0.23.7,<1.0.0", # Latest is 0.23.7 (Oct 16, 2025) - Relaxed to <1.0.0
    "pytest-cov>=5.0.0,<6.0.0", # Latest is 7.0.0 (Sep 9, 2025) - Relaxed to <6.0.0
    "httpx>=0.27.0,<1.0.0", # Matched to runtime version

    # --- QUALITY (Latest Versions) ---
    "mypy==1.18.2", # Latest stable version (Sep 18, 2025)
    "ruff==0.14.2", # Latest stable version (Oct 23, 2025)
    "pre-commit==4.3.0", # Latest stable version (Aug 9, 2025)
    "bandit>=1.8.6,<2.0.0", # Latest stable version is 1.8.6 (Jul 5, 2025)
    "safety>=3.6.2,<4.0.0", # Latest stable version is 3.6.2 (Sep 24, 2025)
]
test = [
    "pytest>=8.2.2,<9.0.0",
    "pytest-asyncio>=0.23.7,<1.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "httpx>=0.27.0,<1.0.0",
    "factory-boy>=3.3.0,<4.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/Harmeet10000/langchain-fastapi-production"
Documentation = "https://github.com/Harmeet10000/langchain-fastapi-production#readme"
Repository = "https://github.com/Harmeet10000/langchain-fastapi-production"
"Bug Tracker" = "https://github.com/Harmeet10000/langchain-fastapi-production/issues"


# Tool configurations
[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.yml", "*.yaml", "*.json"]

# Ruff configuration
[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "ISC001", "TRY003", "PLR0913", "PLR2004"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "PLR2004", "ARG001"]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
known-third-party = ["fastapi", "pydantic", "sqlalchemy", "langchain"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Ty configuration (approximate migration from MyPy)
# Note: Ty does not have 1:1 mappings for all MyPy options, as it uses its own rule-based system.
# Many strict MyPy checks (e.g., disallow_untyped_defs, check_untyped_defs) are covered by default
# "error"-level rules like invalid-return-type, invalid-assignment, etc.
# For per-module overrides like ignore_missing_imports, Ty lacks direct support (file-based globs only).
# Here, we approximate by globally ignoring unresolved-import (common for untyped third-party libs).
# Install type stubs (e.g., types-redis, openai-stubs) where possible for better checking.
# Use # ty: ignore[unresolved-import] on specific import lines if needed for precision.

[tool.ty]
# Equivalent to MyPy's python_version
environment.python-version = "3.12"
# Equivalent to show_error_codes (full output includes codes)
terminal.output-format = "full"
# Optional: Treat warnings as errors for stricter enforcement (no direct MyPy equiv, but aligns with overall strictness)
# terminal.error-on-warning = true

[tool.ty.rules]
# warn_redundant_casts=true → redundant-cast (default is warn, explicit for clarity)
redundant-cast = "warn"
# warn_unused_ignores=true → unused-ignore-comment (default ignore, set to warn)
unused-ignore-comment = "warn"
# warn_no_return=true → Covered by invalid-return-type (default error)
# warn_unreachable=true → No direct rule; Ty may handle via inference (consider terminal.error-on-warning if needed)
# strict_equality=true → Handled in default type checking (e.g., invalid-assignment)
# Approximate ignore_missing_imports for listed modules (global; set unresolved-import to ignore)
# This prevents errors on imports from untyped libs like langchain.*, openai.*, etc.
unresolved-import = "ignore"
# Other disallow_* options (untyped_defs, incomplete_defs, untyped_decorators) → Default "error" rules like
# invalid-declaration, invalid-parameter-default cover similar ground; no change needed.
# warn_return_any=true → No direct; invalid-return-type catches mismatches, including Any issues.
# warn_unused_configs=true → No direct; Ty has no equivalent (config validation is built-in).
# no_implicit_optional=true → Handled in Ty's type inference by default.


# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Bandit configuration (security linting)
[tool.bandit]
exclude_dirs = ["tests", "migrations"]
skips = ["B101", "B601"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.14.2",
    "ty>=0.0.1a25",
]
